Optimization: Code downward motion

+++ src/main/java/org/systemf/compiler/analysis/FrequencyAnalysisResult.java
@@ -5,7 +5,7 @@ import org.systemf.compiler.ir.block.BasicBlock;
 import java.util.Map;
 
 public record FrequencyAnalysisResult(Map<BasicBlock, Integer> frequency) {
-	public long frequency(BasicBlock block) {
+	public int frequency(BasicBlock block) {
 		return frequency.get(block);
 	}
 }
